formatting

+++ test/utils/CommonUtilsLib.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.28;
 import {Vm} from "forge-std/Test.sol";
 
 library CommonUtilsLib {
-    function skip(Vm vm, uint256 time) internal {
+    function skip(Vm vm, uint time) internal {
         vm.warp(vm.getBlockTimestamp() + time);
     }
 }
